Use also catz zones count when calculating zone mem/task pool size
Currently `named` does not take into account the catalog zones' member
zones count when calculating the sizes of the zone mem/task pools.

Create a function in catz.c to calculate the member zones in all catalog
zones in the collection, and use that in `named` when counting the
number of all zones.