Fix non up-to-date comment about zone manager task pool
The comment states that the size of the zone mem/task pools can not be
resized, which is no longer true.

Change the comment to make clear that it is possible to increase the
size of the zone mem/task pools.

Also remove a function requirement which is no longer valid.