Write a function to request relative changes of zone mem/task pool sizes
Currently it is only possible to request zone mem/task pool size change
based on a pre-calculated number of all available zones.

Create a function so that it is possible to request a change by some
amount, without knowing the size of the currently allocated resources.

Also, because it is only possible to increase the size of the pools,
keep track of actually requested size for the cases when the change
is negative, in order to take that decrease into account when requested
to allocate more resources in the future.