Create private keys with GnuTLS PKCS#11 library
If named is started with a specific engine (through -E parameter), and
named.conf has a keystore configured with a PKCS#11 URI, zones that
are using a dnssec-policy that uses such a keystore should create keys
via the PKCS#11 interface. Those keys are generally stored inside an
HSM.

The approach here is to create such keys via the GnuTLS PKCS#11 API,
and then use the 'fromlabel' code to create the DST key, then storing
the required key files on disk.

This requires changes to the build files, to detect the existence of
the GnuTLS library and to add the appropriate flags and libraries to
the make files.

Some changes to the code are required, to store the engine reference
into the keystore.