Manage system test helper scripts by Makefile
When the system test is started from an out-of-source-tree directory, a
specific set of system test helper scripts has been copied from
bin/tests/system in the source directory to the out-of-tree one. The
copying has been done in run.sh and thus for every system test executed
Consequently, those files were never deleted and lingered in the
out-of-tree directory. This is a problem for "distcheck". The copying
mechanism has been moved from run.sh to bin/tests/system/Makefile.am,
and a mechanism that cleans helper scrips was also added.