Remove case_sensitivity from the hash functions
In isc-projects/bind9!6789, the option for case insensitivity was added
to the isc_siphash API, so it can be passed top-down.  In this MR, the
other approach is being used - it's callers responsibility to pass the
downcased key to the the isc_ht API and the option for case
insensitivity has been removed from the isc_ht and isc_siphash API
making them simpler and smaller.

In future, we should handle all dns_names in lowercase internally,
removing the need to lowercase then on demand.