Allow DNSKEY when syncing secure journal
When synchronizing the journal from the unsigned version of the zone
to the secure version of the zone, allow DNSKEY records to be synced,
because these may be added by the user with the sole intent to publish
the record (not used for signing). This may be the case for example
in the multisigner model 2 (RFC 8901).