Collect existing system tests written in Python
Use pytest to collect the existing system tests in Python and execute
them along with shell system tests. Make the system test setup/teardown
into a fixture that can be used for both shell tests and other pytests.

Note that the setup/teardown is executed once for each Python file
(module) and for the tests.sh. The legacy run.sh solution behaves
slightly different. It used to only restart the servers before running
all remaining Python tests.

In order to avoid bending the pytest framework too much, it was chosen
to run entire setup/teardown for each test module (python file /
tests.sh). It doesn't seem to have a significant effect on the total
runtime, but it comes with some drawbacks:

- The system tests within a single directory must be executed
  sequentially and not in paralllel.
- If any of the tests fails, simply following through with remaining
  tests in the same directory would wipe the logs and make failure
  investigation difficult. This is mitigated in the next commit.