Check for log files during system test runs
A single system test directory may contain multiple test cases (tests.sh
and one or more Python tests*.py files). These are executed sequentially
and in between these tests, a complete setup and teardown is performed,
including running clean.sh.

If the tests succeed, clean.sh is called and any logs are removed.
However, if the test fails, logs should be kept in order to facilitate
investigation of the failure.

If there's multiple tests in a single system test directory, a sequence
of failure-success would wipe the logs even though there was a failure.
In order to avoid that issue, check that there are no logs present
before each test. If there are, assume that they are present because a
previous test case has failed and mark the current test as skipped.

This way, if any of the test cases fails, the remaining test cases in
the same system test directory will be skipped and logs will be
available for investigation.