Improve test result handling in pytest
Pytest doesn't expose the test results from individual test cases in
fixtures. Utilize the pytest_runtest_makereport() hook to put the test
results into session, which is accessible everywhere.

There is also a need to aggregate the individual test results into a
single outcome. This is needed because the setup/teardown of servers is
module-wide, but there might be multiple individual test results. A
single result is needed to decide whether the test results can be
cleaned up or not.

Finally, reorganize the code to better utilize the pytest.fail() and
pytest.skip() mechanism. When these are called in fixtures, code
execution is ended with the exception of finally blocks.