switch to using isc_loopmgr_pause() instead of task exclusive
change functions using isc_taskmgr_beginexclusive() to use
isc_loopmgr_pause() instead.

functions that were implemented as task events because they needed to be
running in a task to use exclusive mode can now be changed into loop
callbacks instead. (the exception is catz, which is being changed in
a separate commit because it's a particularly complex change.)