pytest: obtain env vars from conf.sh
The commands executed by pytest during a system test need to have the
same variables defined as if they were executed by the run.sh shell
script.

It was decided that for the moment, legacy way of executing system tests
with run.sh should be kept, which complicates things a bit. In order to
avoid duplicating the required variables in both conf.sh and pytest, it
was decided to use the existing conf.sh as the only authoritative
place for the variables.

To access these variables in Python, the env command is utilized to list
the variables and parse them. Note that the variables need to be
explicitly exported, because there is no POSIX-compliant portable way to
obtain a list of non-exported shell variables.

This solution is a bit hacky and is only meant to be used during the
transitory phase when both pytest and the legacy run.sh are both
supported. In the future, a superior pytest-only solution should be
used.

For discussion of other options, refer to
https://gitlab.isc.org/isc-projects/bind9/-/merge_requests/6809#note_318889