Add --clean option to pytest
Allow user to skip the check for test artifacts and clean them up
instead. This may be useful for re-running failed tests without having
to manually call clean.sh.

This option isn't and can't be set as default. The reason is this
mechanism is used as a detection of previously failed tests when
multiple tests suites (e.g. tests.sh and tests_x.py) are contained
within a single directory to avoid re-writing the logs necessary for
debugging.

This option should only be used with caution when debugging/developing a
single test. Avoid using this option when multiple test suites are
involved.