Allow port to be combined with address in config
Allow configuring and address with port appended with 'p' delimiter,
for example:

    query-source: 2001:db8::1p80;

This can already be configured as follows:

    query-source: 2001:db8::1 port 80;

But can become messy if we want to configure multiple addresses on a
single line that also have 'port' option:

    primaries port 10053 source 10.53.0.1 port 10053 source-v6 2001:db8::1 port 10053 {
    ...;
    }

vs:

    primaries port 10053 source 10.53.0.1p10053 source-v6 2001:db8::1p10053 {
    ...;
    }

----

Introduce a new flag CFG_ADDR_PORTOK that tells if an address can be
followed with a 'port' option (similar to CFG_ADDR_DSCPOK for 'dscp').

Split the 'parse_querysource' function to make use of a helper function
that we also are going to need for parsing source addresses that don't
allow following 'port' and 'dscp' options.

----

There are many different ways to combine addresses and port numbers
that are represented in text. Examples are shown below and taken from
RFC 5952.

- [2001:db8::1]:80
- 2001:db8::1:80
- 2001:db8::1.80
- 2001:db8::1 port 80
- 2001:db8::1p80
- 2001:db8::1#80

We actually use the latter (with '#' delimiter) already in our logs, but
since that is also a character that starts a comment in our named.conf
files, we cannot use it as port delimiter. Easiest alternative is the
'p' delimiter.

The RFC says the first one (with the square brackets [])
should be employed and should be the default. We could change our
isc_sockaddr_totext() and parse functions to take this format, at the
cost of some additional parsing code complexity.