Use common name convention for system tests
Don't use underscore or dash in system test names and always use
lowercases without spaces for the names even when the test name consists
of multiple words.

Before this change, there were three different ways of naming the tests:
as described above, or with underscore/dash in between the words.